feat: add endpoint summary

+++ internal/repository/kategori.go
@@ -52,6 +52,6 @@ func (r kategoriRepo) Update(ctx context.Context, kategori *model.Kategori) erro
 	return err
 }
 func (r kategoriRepo) Delete(ctx context.Context, kategoriId int) error {
-	err := r.db.DB().Model(model.Kategori{ID: kategoriId}).Delete()
+	err := r.db.DB().Model(model.Kategori{ID: int32(kategoriId)}).Delete()
 	return err
 }
